Connector util works on Windows.
