+++ gcp-connector-util/common.go
@@ -1,0 +1,9 @@
+/*
+Copyright 2015 Google Inc. All rights reserved.
+
+Use of this source code is governed by a BSD-style
+license that can be found in the LICENSE file or at
+https://developers.google.com/open-source/licenses/bsd
+*/
+
+package main
