+++ lib/config_unix.go
@@ -10,6 +10,47 @@ package lib
 
 const defaultConfigFilename = "gcp-cups-connector.config.json"
 
+type Config struct {
+	commonConfig
+
+	// CUPS only: Where to place log file.
+	LogFileName string `json:"log_file_name"`
+
+	// CUPS only: Maximum log file size.
+	LogFileMaxMegabytes uint `json:"log_file_max_megabytes"`
+
+	// CUPS only: Maximum log file quantity.
+	LogMaxFiles uint `json:"log_max_files"`
+
+	// CUPS only: Log to the systemd journal instead of to files?
+	LogToJournal bool `json:"log_to_journal"`
+
+	// CUPS only: Filename of unix socket for connector-check to talk to connector.
+	MonitorSocketFilename string `json:"monitor_socket_filename"`
+
+	// CUPS only: Maximum quantity of open CUPS connections.
+	CUPSMaxConnections uint `json:"cups_max_connections,omitempty"`
+
+	// CUPS only: timeout for opening a new connection.
+	CUPSConnectTimeout string `json:"cups_connect_timeout,omitempty"`
+
+	// CUPS only: printer attributes to copy to GCP.
+	CUPSPrinterAttributes []string `json:"cups_printer_attributes,omitempty"`
+
+	// CUPS only: use the full username (joe@example.com) in CUPS job.
+	CUPSJobFullUsername bool `json:"cups_job_full_username,omitempty"`
+
+	// CUPS only: ignore printers with make/model 'Local Raw Printer'.
+	CUPSIgnoreRawPrinters bool `json:"cups_ignore_raw_printers"`
+
+	// CUPS only: ignore printers with make/model 'Local Printer Class'.
+	CUPSIgnoreClassPrinters bool `json:"cups_ignore_class_printers"`
+
+	// CUPS only: copy the CUPS printer's printer-info attribute to the GCP printer's defaultDisplayName.
+	// TODO: rename with cups_ prefix
+	CUPSCopyPrinterInfoToDisplayName bool `json:"copy_printer_info_to_display_name,omitempty"`
+}
+
 // DefaultConfig represents reasonable default values for Config fields.
 // Omitted Config fields are omitted on purpose; they are unique per
 // connector instance.
